minor fixes in tests and dynamic

+++ PySDM/dynamics/relaxed_velocity.py
@@ -1,4 +1,4 @@
-import numpy as np
+from math import exp
 
 from PySDM.attributes.impl.attribute import Attribute
 from PySDM.particulator import Particulator
@@ -29,7 +29,7 @@ class RelaxedVelocity:  # pylint: disable=too-many-instance-attributes
         self.tmp_data.product(self.terminal_vel_attr.get(), self.volume_attr.get())
         self.tmp_data *= self.rho_w
         self.tmp_data -= self.fall_momentum_attr.get()
-        self.tmp_data *= float(1 - np.exp(-self.particulator.dt / self.tau))
+        self.tmp_data *= 1 - exp(-self.particulator.dt / self.tau)
 
         self.fall_momentum_attr.data += self.tmp_data
